+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Enh #4: Add trait `HasFirstItemClass` class (@terabytesoftw)
 - Enh #5: Add trait `HasFirstLinkClass` class (@terabytesoftw)
 - Enh #6: Add trait `HasLastItemClass` class (@terabytesoftw)
+- Enh #7: Add trait `HasLastLinkClass` class (@terabytesoftw)
 
 ## 0.1.0 March 5, 2024